Starting to refactor test, adding http post to wallet sender
